requirements.txt
rich
PyYAML
argparse
requests
